r16284: Start fixing up gcc4 -O6 warnings on an x86_64 box. size_t != unsigned int in a format string. Jeremy. (This used to be commit face01ef01e1a3c96eae17c56cadf01020d4cb46)

+++ source3/nsswitch/winbindd_cache.c
@@ -229,8 +229,8 @@ static time_t centry_time(struct cache_entry *centry)
 {
 	time_t ret;
 	if (centry->len - centry->ofs < sizeof(time_t)) {
-		DEBUG(0,("centry corruption? needed %d bytes, have %d\n", 
-			 sizeof(time_t), centry->len - centry->ofs));
+		DEBUG(0,("centry corruption? needed %u bytes, have %u\n", 
+			 (unsigned int)sizeof(time_t), (unsigned int)(centry->len - centry->ofs)));
 		smb_panic("centry_time");
 	}
 	ret = IVAL(centry->data, centry->ofs); /* FIXME: correct ? */
